Only build tarball once for client libs.
For any given revision, we should only ever create one tarball.
Similarly, we should upload that tarball and only that tarball to
tarballs.openstack.org as well as pypi if it's appropriate to do so.

Library projects are sufficiently different at this point that I made a
different template for them. It would be neat if we could get nested
templates.

Change-Id: Iec81d0d5c0c3a4a8021213fe4bb6884c45fb483a