Create new coverage job for heat.
The old coverage job for heat was used in the check queue, but that job
was configured for being in the zuul post pipeline so the copying of
logs failed. Correct this by creating a new job that is configured for
the proper pipeline.

Change-Id: Iec54fd463e2535660dcf48ff3408fd8d0c2eb10d
Reviewed-on: https://review.openstack.org/13204
Reviewed-by: James E. Blair <corvus@inaugust.com>
Approved: Clark Boylan <clark.boylan@gmail.com>
Reviewed-by: Clark Boylan <clark.boylan@gmail.com>
Tested-by: Jenkins