zuul can now be installed from a different git repo
$git_source_repo is a new optional parameter to the zuul class, let you
specify from which repository you want to install zuul. The default is
the maintainer version on GitHub.

The reason for this change is that some third party always hosts their
software on their own infrastructure. That is the case of the Wikimedia
Foundation which is reusing this module.

Change-Id: I5ac8852b08d158ff3531f6795855df20a204789e
Reviewed-on: https://review.openstack.org/13576
Reviewed-by: Clark Boylan <clark.boylan@gmail.com>
Approved: Monty Taylor <mordred@inaugust.com>
Reviewed-by: Monty Taylor <mordred@inaugust.com>
Tested-by: Jenkins