Ignore devnull branch for post jobs.
This is a temporary measure to allow ttx to create a branch which
does not publish tarballs in order to merge the version number
change for stable/folsom.  In the future, we should allow
zuul to be configured to ignore branch creation and deletion for
certain branches.

Change-Id: I071604da1f70453123826d29256beb7a54e5f31c
Reviewed-on: https://review.openstack.org/13712
Reviewed-by: Monty Taylor <mordred@inaugust.com>
Reviewed-by: Clark Boylan <clark.boylan@gmail.com>
Approved: James E. Blair <corvus@inaugust.com>
Tested-by: Jenkins