Add saltstack puppet module.
Add a puppet module to install saltstack minion and saltstack master.
This is to be used to bootstrap saltstack. As a result the only
configurable option here is the master node that minions should talk to.
Eventually salt master should have a setup similar to puppet to manage
its own configs.

Change-Id: I27cb902f043c4f4a2a53c599c08ae94df297cc97
Reviewed-on: https://review.openstack.org/14235
Reviewed-by: Monty Taylor <mordred@inaugust.com>
Reviewed-by: Paul Belanger <paul.belanger@polybeacon.com>
Approved: James E. Blair <corvus@inaugust.com>
Reviewed-by: James E. Blair <corvus@inaugust.com>
Tested-by: Jenkins