Check the number of tests that were run after tox before success
This adds a dirty check of the output file after a tox run to make
sure that at least one test passed. If not, then we shouldn't
allow a successful result. This is intended to catch the case where
none of the tests run because of some parse or test setup error,
which might otherwise be considered a successful run.

Change-Id: I598262f47971e238876b38416c00a70de8313543
Reviewed-on: https://review.openstack.org/14402
Reviewed-by: Monty Taylor <mordred@inaugust.com>
Reviewed-by: Clark Boylan <clark.boylan@gmail.com>
Approved: James E. Blair <corvus@inaugust.com>
Reviewed-by: James E. Blair <corvus@inaugust.com>
Tested-by: Jenkins