Correct stackforge .gitreview docs.
The manage projects script adds a gitreview file for you if the new
projects does not have an upstream. Make it clear that you only need to
add the .gitreview file on projects with an existing upstream.

Also, fix a tiny typo in the Jenkins document.

Change-Id: I74c07d9cd952509bd9ddc883a833cb3e22b13b38
Reviewed-on: https://review.openstack.org/18570
Reviewed-by: Monty Taylor <mordred@inaugust.com>
Reviewed-by: James E. Blair <corvus@inaugust.com>
Approved: James E. Blair <corvus@inaugust.com>
Tested-by: Jenkins