Gate tempest on full tempest + postgres.
There was a change to have the test env decide if tempest should run all
of its tests instead of having devstack do this. This change made the
devstack postgres test run only the tempest smoketests when tempest was
being gated. Add a non voting postgres-full job that will run in the
tempest project's gate to change the test behavior back to what it was.

Change-Id: Ifaaef22437a2582ff7d8c3714d577ff42bb8db7d
Reviewed-on: https://review.openstack.org/20068
Reviewed-by: Jeremy Stanley <fungi@yuggoth.org>
Reviewed-by: Sean Dague <sdague@linux.vnet.ibm.com>
Approved: James E. Blair <corvus@inaugust.com>
Reviewed-by: James E. Blair <corvus@inaugust.com>
Tested-by: Jenkins