Work around tox+rhel+distribute+setup_requires.
Trying to use setup_requires with the version of distribute on rhel
doesn't work because that version doesn't properly re-import itself.
When tox gets involved, it tries to run setup.py sdist outside of
the created virtualenv, which means it's ignoring our declared deps
for that action. The bug in distribute is worked around by running
any setup.py command once and letting it fail, then running again.

So - python setup.py --version is the cheapest setup.py command to
trigger the behavior that needs to happen. Throw the error results to
/dev/null because we expect there to be a traceback and we don't
want to confuse people reading log outputs. Leave stdout because
it's good to be able to track which setup_requires requirements
it is pulling down before it dies on not actually being able to
use them.

Change-Id: I982a08ffa5784ed97f70f41d91a58da66b7498b6
Reviewed-on: https://review.openstack.org/26533
Reviewed-by: James E. Blair <corvus@inaugust.com>
Approved: Clark Boylan <clark.boylan@gmail.com>
Reviewed-by: Clark Boylan <clark.boylan@gmail.com>
Tested-by: Jenkins