Change gate-*-docs success URL.
Drill down all the way to the docs.  As long as this job is only
running in the gate context (it is), run-docs.sh won't do anything
funny like copy or move the resulting docs around, so it should
be safe to link to doc/build/html.

Change-Id: I1c8aaed9020b7c9872ff0ccc876bf719df4cd5dd
Reviewed-on: https://review.openstack.org/26643
Reviewed-by: Jeremy Stanley <fungi@yuggoth.org>
Reviewed-by: Clark Boylan <clark.boylan@gmail.com>
Approved: James E. Blair <corvus@inaugust.com>
Tested-by: Jenkins