Puppet dashboard mysqld innodb_file_per_table.
In 232747f we attempted to implement it this way initially, but
subsequent patchsets went off into the weeds instead trying to use
mysql::server::config which, as it turns out, was not introduced
until a newer version of the puppetlabs mysql module than we use.
Rather than go through a potentially disruptive infrastructure-wide
upgrade right this moment, this patch gets the job done in a more
traditional fashion.

Change-Id: I6967388072c7c8b3e170048e23618ea95bccfecc
Reviewed-on: https://review.openstack.org/27511
Reviewed-by: James E. Blair <corvus@inaugust.com>
Reviewed-by: Anita Kuno <anita.kuno@enovance.com>
Reviewed-by: Khai Do <zaro0508@gmail.com>
Reviewed-by: Paul Belanger <paul.belanger@polybeacon.com>
Approved: Clark Boylan <clark.boylan@gmail.com>
Reviewed-by: Clark Boylan <clark.boylan@gmail.com>
Tested-by: Jenkins