Corrected IRC channel for stackforge/cookbook-openstack-*
We actually watch #openstack-chef, not #chef-openstack.
The current channel is incorrect, and it wasn't until I mentioned it in #openstack-chef that we noticed.
Only openstackgerrit was in that channel. The entire 'openstack-chef' community is in #openstack-chef.
This channel has 35 people in it (currently) whereas the other one has 3.

Change-Id: I4f5b64a36a8b6e1dd3a404f005bbac09d784d134
Reviewed-on: https://review.openstack.org/29315
Reviewed-by: John Dewey <john@dewey.ws>
Reviewed-by: James E. Blair <corvus@inaugust.com>
Reviewed-by: Anita Kuno <anita.kuno@enovance.com>
Reviewed-by: Elizabeth Krumbach Joseph <lyz@princessleia.com>
Approved: Jeremy Stanley <fungi@yuggoth.org>
Reviewed-by: Jeremy Stanley <fungi@yuggoth.org>
Tested-by: Jenkins