Fix periodic-tempest-devstack-vm-stress job
The job was not actually running the stress tests, it would exit after
the devstack exercises because the DEVSTACK_GATE_TEMPEST variable was
not set. This variable is required to run any tempest tox job. This
commit corrects the behavior by setting DEVSTACK_GATE_TEMPEST to 1.

Change-Id: Ica2100a43b86cb4b11d090b39d9f98e4e8bbc224
Reviewed-on: https://review.openstack.org/37099
Reviewed-by: Jeremy Stanley <fungi@yuggoth.org>
Approved: Clark Boylan <clark.boylan@gmail.com>
Reviewed-by: Clark Boylan <clark.boylan@gmail.com>
Tested-by: Jenkins