Mark puppet-lint jobs as voting
Since the introduction of non-voting puppet-lint jobs for OpenStack
puppet modules, a lot of cleanup have been done in the manifests.

All puppet-lint ERRORS have been fixed and builds are now passing
successfully. We wish to enable voting for puppet-lint jobs.

We are aware that WARNINGS are still present but won't fail the build.
Fixing WARNINGS will be a work in progress in the following months.

Change-Id: I155154b7a3fe2c18130bdcb6b76be11f57c0a0ed