Adds support to specify a different jenkis_jobs_builder username
Add the possibility to choose another username for jenkins_jobs_builder script.
If you use another authentication method (LDAP for example), and the username
is an email, the jenkins_job_builder fails, since gerrig isn't a valid user in
the LDAP directory. This won't break the actual configuration. If the user don't
set the jenkins_jobs_username, it will continue to use gerrig as default.

Change-Id: I990cf15c4c583d3041725e8c8f1bd9beb29e0d0f