Update git.o.o configs for better rolling reboots
* modules/openstack_project/manifests/git.pp: Give the haproxy stats
socket admin level permissions. Make the socket owned by user root and
group root with mode of 0600. This is necessary to provide instructions
to enable and disable backend servers through the stats socket. Also,
install socat on this server so that commands can be given through the
stats socket.

* modules/openstack_project/manifests/git.pp: Enable git-daemon service
so that it starts running at boot.

Change-Id: Ic38114bdbd9fd1e9ab711e636deaadb15f9c7fe3