Add stable/havana jobs.
This commit adds stable/havana jobs to JJB and Zuul. These are both
periodic bitrot jobs and check jobs run against devstack-gate. As part
of this process the stable jobs were pulled into a separate file so that
they can be templated across branches. This should simplify the addition
of IceHouse in the future.

Change-Id: I46cab2220153f82ff89a1df0f083a99f9d5e6260