Create and upload wheels
We want to make wheels for pre-release and release events. We will only
build wheels on python 2.7. If the project supports python 3 as well, the
project will have marked itself as supporting universal wheels. If it does
that, python 2.7 will build a wheel that can be used in both places. If it
does not support python 3, then we don't have to build on on python 3.

Change-Id: I060b59d8adc586ff37065fa4f6172a592ee3f903