Drop the saz-gearman module (we don't use it)
This module confused me in that I though we were actually
using gearmand somewhere. Infra has its own version of
gearman fork in python... so no need to keep this around anymore.

This change adds a new remove_module function to install_modules.sh
and updates the script to use that function to remove gearman
and the associated dependency modules.

Change-Id: Ifc3808407eba9853b6991ef9744adb96453d46a8