Consolidate rsyslog configs.
Puppet doesn't allow non unique namevars. Having mutliple rsyslog
service definitions violates this rule. Work around this by
consolidating all rsyslog configs in openstack_project::template.
In the process of doing this it became apparent that the ensure whoopsie
was absent resource should be consolidated in to template.pp as well so
this was done as part of the larger consolidation.

Change-Id: I7effed7bacb85386f0cf36edffe442e2cad944d0