Turn on devstack exercises for neutron (non-voting)
Currently devstack exercises are not working for havana neutron.
This makes it very difficult to write the required grenade job
for neutron.

This patch turns devstack exercises on for neutron in check,
jobs enabling devstack exercise failures to be tracked.
There is a provision to add this job to periodic jobs in the
future.

Change-Id: Id177bc58cdaf78dddece1ac1ac137cc03d32bd9c