Explicitly document requirements for 3rd party testing
As we are moving towards requiring more and more third party testing, we
should make sure we have a central and explicit list of requirements.
These are the requirements to have a voting third party testing system,
not the requirements of what the system should be running (to count a
driver as fully tested).

Change-Id: Ic1e1cdf9af9cd21e29a0e768d5d78ae4f84108fb