provide time in queue in zuul ui
this change adds the time in the queue to the zuul ui, to make it
much easier to understand how backed up we are at the moment. A
judgment call is made that changes over 2hrs in queue get their
times turned yellow, and changes over 4hrs in the queue get their
times turned red.

this makes the header now 2 lines of output, and some padding
fixes were applied to make the compact version not have the ugly
white underbar in them.

For periodic jobs, we need to force a &nbsp; to get the block
layout to hold.

Change-Id: I7ca65b249aea65b2d93e2330d5f4ee20cdb913fd