Add build job for Japanese Install Guide
We now have a 100% translated Japanese install guide good to go.
This patch adds a new install-guide build job for foreign languages,
done in a generic way so future foreign languages can be easily
added. It adds Japanese language build jobs, since this is the
first fully translated language.

In addition, the opportunity was taken to genericise the
openstack-operations-guide-ja build job so it can be used
for languages other than Japanese.

NB: this is a first time working with templated jobs and variables,
appreciate careful review ... especially don't want to break the
openstack-operations-guide-ja job as I just got that working again!

Change-Id: I8a1c6462e6a136e05d4313dbfc6dc600d8395b79