Ensure a recent (48h) check result
Add a comment-added filter to the check pipeline that matches
whenever there is an existing vote from jenkins that is more
than 48 hours old.

This should mean that any change that has activity on it (in the
form of anyone or any bot leaving comments) will have a relatively
recent check result.  This value may be changed later, but I'm
setting it to a moderately conservative 48h intially to hopefully
reduce the potential rush.

Change-Id: I814660aead8ccc8c089594775fe7de62b63f3903