Fix check_projects_yaml_alphabetized.sh
The previous patch moved " PROJECTS_LIST=$OLDPWD/modules/..." before
the pushd call. But pushd changes the path and thus OLDPWD as well.

So, there are two possible fixes: Either remove the OLDPWD variable from
the if and set PROJECTS_LIST in another way with a full path or
move the pushd call before the if.

Change-Id: If47ab0cfb0078d5e570057c974e3d1f0367d4e0a