Remove taskflow sqlalchemy 0.9 testing
Until further notice, after the switch to using
the openstack pypi mirror (which does not mirror 0.9)
we need to turn off the venv tests that probe & verify the
compatability with sqlalchemy 0.9.

This is planned to be fixed but is currently a blocker
for taskflow unit tests and gating, since it causes a
dependency resolution failure when testing (dependency
not found).

Partially fixes bug 1280008

Change-Id: I07f6168d192ce0f60d7ef14b107b802d0fe50d54