change 3 hr jobs to 2 hrs
we've never seen a valid job exceed 2 hrs, and the inner timeout
calculation is now working correctly to ensure that we always get
log results. It's thus safe to go back to a 2hr timeout, which
will also help when jobs actually go off the rails in a bad way
and drive all the way to a full timeout.

Change-Id: If69b3cc41bda6689da285fcd652570745ae3713a