Fix Fedora Puppet failure to use pip
In Fedora the pip command used by the packaged python-pip is
"pip-python" rather than "pip" due to a naming conflict. The
Puppet module tries to compensate, but we don't use the Fedora
package, so the command really is "pip" for us.

Adding symlink line to install_puppet.sh for Fedora to add the
pip-python command that Puppet is looking for.

Change-Id: I60a03fb6c27a18cd8a536080bcce1baaba965237