Add gate-magnetodb-devstack job
Add an initial devstack job for the MagnetoDB project.
It just installs magnetodb's devstack integration and runs devstack.
That just ensures that the devstack integration doesn't blow up just
trying to start up the service(s). There's not much else to test yet,
but this is a starting point to build on.

Start this job out as non-voting so that we can make sure it works as
expected before turning it on to be gating.

Change-Id: Ie99a88b95b28cee9443b05fd091f9e4f590e4cfb