Test d-g against partial n-cpu job.
We made a change to d-g that moved grenade testing forward from
havana -> master to icehouse -> master. This worked fine for the normal
grenade jobs but not the grenade partial n cpu job. To make sure we test
this in the future gate d-g on the partial n cpu job.

Change-Id: I67d89167d3fbbaf21d54f5d8c1bcc6c921a49cbb