Check for new files as well as changed files
Currently, the run-compare-xml.sh script will only report changed or
removed files, it does not report new files being created. This change
replaces diff in a for loop with a recursive diff that treats missing
files as empty files. The result is that the output is presented as a
single unified diff that includes new files as well as changed and
removed files.

Change-Id: I0bb5a1ee93e689666ca5080ec5ccbe036c73d682