Fix scenario name in rally-neutron job
Currently there is two jobs in rally group: dsvm and dsvm-neutron.
Scenario name should be different for this jobs: {name} for dsvm
and {name}-neutron for dsvm-neutron. E.g if we want to test glance
with and without neutron, we should create two scenario files in
rally-scenarios directory: glance.yaml and glance-neutron.yaml

Change-Id: Ic7451e8e6b0b06c136db0f15ab71ca51e26dd67a