disable voting on gate-tempest-dsvm-virtual-ironic
This job is passing only 66% of the time in the gate, 30 fails in
the last 48hrs. That's too low to be allowed to vote in the gate.

Due to the oslotest overlap, failing on this configuration now
impacts all of OpenStack. Some initial analysis shows that this
configuration is doing *a lot* more interaction with the network
apt servers. That's something which fundamentally needs to be
limitted.

Instead this job is going to need to work on building a new nodepool
image that includes the cloud archive, instead of installing it
on the fly.

This leaves the check job voting, but will not impact the rest of
OpenStack when ironic or diskimagebuilder fail in the gate.

Change-Id: I2652b777e4e24034f715dfecb71a87f128f14644