heat slow voting again
Most of the heat failure happened when the subnet
used by a test case 10.0.3.0/24 is overlapped with
the host's subnet.

The heat-slow job always failed, when the test runner
machine's ip was from 10.0.3.0/24.

The heat/neutron test case now uses the configured subnet,
instead of the hard coded 10.0.3.0/24.

The heat-slow job failures reduced
 from ~90/day to ~3/day.

Change-Id: I428b1d2a2a0de93242af2f89f8bb487c931f304f