Align exim module parameter name with the tree
The rest of the tree, literally everwhere, calls this value
sysadmins - but the exim module calls it sysadmin. This doesn't
really matter of course, but I noticed it while looking in to a
review on another patch and couldn't not fix it. Also, there are
two comments that are clearly remnants from having used exim.pp
as a starting point.

Change-Id: I184701c8c4d4b1e9d93d746b2ad0c5fc53443866