Don't auto-recheck when last check was false
When Jenkin's vote is 72 hours old it'll re-run checks on the next
comment to ensure that its vote is still valid. However this is only
necessary to confirm that postive votes are valid. If Jenkins
originally voted -1 then there is no point in rechecking (this can
be done through the 'recheck bug xyz' filter)

Change-Id: Ia24b09df79c33f425763d8b03bf58707f8fc1046