Clarifies that CI accounts must request accounts from infra
Currently there is ambiguity in the third party documentation
around requesting accounts. Some have interpreted the documentation
as stating it is possible to conduct tests using a gerrit user
account. This patch addresses the documentation language making
it clear that a third party ci system is required for third party
ci interactions with gerrit and the system must be created by
infra.

This patch also changes the recommended method of requesting a
third party ci account from three choices (email, bug or irc) to
one choice (email). This edit was previously part of
https://review.openstack.org/#/c/101013/ and has been moved into
this patch.

Change-Id: Iab88b729d714dcaa881376e9237834e877acecfa