Add pre-requisites needed for diskimage-builder
For nodepool to use diskimage-builder, it needs a set of pre-reqs
on the nodepool host. Install them.

Also create /opt/nodepool_dib for building images.  This is
deliberately placed on the non-ephemeral disk as new host builds have
a large partion mounted here.  Testing shows test images were coming
in at about 2.6GB each. If we figure two images for precise and trusty
and one each for centos6 and f20 that is about 15GB.

Change-Id: I725cad999f795a556fd89818578b78b9114fb110