Add gating for security-doc repo
The security-guide has moved from openstack-manuals repo to the
security-doc repository.

Add the usual manual gating jobs for the security-doc repository.
The guide is translated into Japanese, so publish the Japanese
document as well. This needs a new job 'security-doc-{lang}'.

Add all jobs to Zuul for execution.

Change also translation scripts, the security-guide document is on
top-level of the repository and not below doc. Thus, we need to special
case it.

Change-Id: Ib267a35566d446fa13675e4c7247222916f6ffe4