Update partially translated glossary
Download partially translated glossary the same way we handle the common
directory, this is good enough for using the glossary as part of manuals.

Refactor the code so that we do not invoke tx pull multiple times (first
for all files, then for the special ones): We now setup the translations
with minimal translation in them and update as appropriate.

Also, skip glossary project in all repos besides openstack-manuals.

Change-Id: Ia50c3a76b0498c85849fa7895fd1d405e5915464