Stop using select-mirror in slave scripts
It turns out that we pass a lot of args around simply for the
purpose of doing mirror selection. This goes away if everything
gets a mirror.

This change should be merged and then nodepool images updated before
proceeding with the child change removing use of mirror selection
parameters in jobs.

Co-Authored-By: Monty Taylor <mordred@inaugust.com>
Change-Id: I9198db5570631dcdea18b645beee27322e6f05f6