convert to two level recheckwatchbot.yaml
As of I7353af4c3f141d4bd617d6fd388d7957e0586ba8 we'll introduce
the ability to have a 2 layer config which includes the messages
that we'll send back to the user on gerrit outside of the
elastic-recheck code base. This makes it easier to make changes
to the UX consistent with what we want to change with the rest
of the config.

Change-Id: I9208123a4cb3be02c272cd8a6eba460f4130a960
Depends-On: I7353af4c3f141d4bd617d6fd388d7957e0586ba8