remove recheck 'no bug' & 'bug #'
At the QA/Infra sprint we decided that we're getting pretty
minimal value from asking people for bugs # in the tool, as we're
not really using that information. Instead, just convert to a
'recheck' string, which can include an arbitrary comment after it,
so that if people want to provide us with info, we can get it
out later, but it's no longer mandatory.

Change-Id: I419b0edf2e1805adbdba83b2630b087c7eaf14d9