Run sqa-migrate py26 jobs on centos6 nodes.
The sqlalchemy migrate special env job for python 26 name does not match
the centos6 assiginment in the zuul openstack functions for node type.
This causes the jobs to end up as not registered with gearman.

Update the regex in the zuul functions to match
gate-sqlalchemy-migrate-26fooblah. This will make the jobs register with
gearman so that zuul can run them.

Fixes-bug: 1347068
Change-Id: I5944889d53656395c98bab9a97cb4e76c793c9fd