build tarballs and docs for barbican
While going through juno-2, I hit a snag for barbican as the tarball
job didn't run.  Update the zuul layout for barbican to match other
parojects for the pre-release and release pipelines.  Also switch to
using the python-jobs template to simplify the barbican config, as
well as include the docs job in the gate.

Change-Id: I0d7294f749112e8f3c67a05cefd7bc3c4901bcb7