Set default test node to trusty for all projects.
Switch the default test node type for all projects to trusty. Any tests
for icehouse and older branches will continue to default to precise.
Py26 will still run on centos, f20 jobs on f20 and so on. This only
changes the default for tests that are not picky about the node they run
on.

Continue to run infra config puppet jobs on precise as most of the infra
nodes are still precise.

Change-Id: Ie794fd250628d3879973647f42b945a36f070471