Enable puppet3 master bootstrap
This enables install_puppet.sh to run successfully on
trusty with PUPPET_VERSION=3 set.

Add a util directory and a puppetmaster bootstrap
shell script.

Puppetdb-terminus versions are the same in trusty and
precise but are different than the puppet version number.

Also puppetdb-terminus isn't present in the puppetlabs
repositories.

Along with this I have done sufficient testing to use
puppet version 3.6 with confidence.

Work has been done to enable directory environments as well.

Change-Id: I925fa6ac1ed9fc9c0b5e9e616fb071bf7a815436