Add integration apply test
Make module installation conditional in install_modules.sh so that
we can specify that we _do not_ want it to install the modules
that will be participating in the integration test.

Add a variant of the apply test script (currently test.sh) that
uses zuul-cloner to check out the appropriate versions of modules
that are participating in the integration test.

Add jobs that run this to the experimental pipeline.

Change-Id: I16055cdd45f6e05aa8ebceb500c5709f265268df