Improve filtering of translations
Add new test:
Do not add new empty files without any translations or messages.
Those are not needed.

Enhance test:
When filtering translations and checking for only changes to .pot files,
check first that a file was really changed.

Otherwise, the .po files might have just timestamp changes and get
filtered out and we end with a patch with only the pot file in it.

Change-Id: I667b76baa5d779488098e1a297f38ea66e53fb6d